alter handling of nullish errors in individual stock components
discussion:

the portfolio reducer  is structured to key into the portfolio in order to populate the selected portfolio item.

line 110 of reducer_for_Portfolio

"const nextSelect = state.portfolio[action.payload]"

This means that that part of the store is actually a pointer to an object in memory

React-Redux is structured to append, essentially, listeners in the componentDidUpdate lifecycle method. The component
is subscribed to changes in the part of the store thats defined as a props- map state to props.

These individual components need to display parts of the selected portfolio item, but their listener only updates
when the selectedPortfolio item itself changes- its just a pointer, NOT when the values in the object that its pointing
to alter.

This means that, if the quantity of the holding changes, map state to props won't trigger a set state call.

My first solution was simply to key further into the selectedPortfolio item object- a pointer directly to the value
quantity.

The issue was, it seems, the React-Redux library attempts to resolve these variable bindings before React renders
the component. The application was throwing nullish errors- "can't find property quantity of undefined" during
runtime, presumably, because the component was updated before or while the action to switch the
selected port item was dispatched.

The solution is the same as the socket listening pattern- to have the component track the variable locally as
part of state, and manually do the same thing basically that React-Redux library does- listen to a prop and
call setState if the prop changes- in componentDidUpdate.

In future, another solution could be to deep copy the selected portfolio item, or else destructure that
object into state, basically hoist the values globally so that mapStateToProps can subscribe to them
directly.